perf(game): 优化游戏性能和精度
- 将子弹和坦克的速度数据类型从Integer改为Double以提高移动精度
- 调整子弹发射速度从18降至9.0以平衡游戏体验
- 修改地图格子尺寸从32px调整为25px以优化渲染性能
- 更新玩家初始速度从6调整为3.0以匹配新的速度系统
- 修复方向和速度计算方法中的参数类型以支持双精度运算

+++ src/main/java/com/zunf/tankbattletcpserver/model/bo/BulletBO.java
@@ -40,7 +40,7 @@ public class BulletBO implements Serializable {
     /**
      * 速度 px/tick
      */
-    private Integer speed;
+    private Double speed;
 
     /**
      * 伤害
